do some funky refactoring

+++ cmd/reader/types.go
@@ -4,5 +4,5 @@ type SuggestResult struct {
 	Prefix string   `json:"prefix"`
 	Word   string   `json:"match"`
 	Other  []string `json:"other"`
-	Hits   int      `json:"hits"`
+	Hits   uint64   `json:"hits"`
 }
